+++ CHANGELOG.md
@@ -15,6 +15,8 @@ Possible log types:
 
 - [changed] Minimal required Rust version 1.20.0 (#68)
 - [changed] Update `log` crate to version 0.4
+- [changed] Use an enum (`IssueReportChannel`) for the `issue_report_channels`
+  field (#4)
 
 ### v0.5.1 (2017-06-17)